+++ lib/simplecov_mcp/commands/command_factory.rb
@@ -7,6 +7,7 @@ require_relative 'summary_command'
 require_relative 'raw_command'
 require_relative 'uncovered_command'
 require_relative 'detailed_command'
+require_relative 'total_command'
 
 module SimpleCovMcp
   module Commands
@@ -17,14 +18,15 @@ module SimpleCovMcp
         'summary' => SummaryCommand,
         'raw' => RawCommand,
         'uncovered' => UncoveredCommand,
-        'detailed' => DetailedCommand
+        'detailed' => DetailedCommand,
+        'total' => TotalCommand
       }.freeze
 
       def self.create(command_name, cli_context)
         command_class = COMMAND_MAP[command_name]
         unless command_class
           raise UsageError.for_subcommand(
-            'list | summary <path> | raw <path> | uncovered <path> | detailed <path> | version'
+            'list | summary <path> | raw <path> | uncovered <path> | detailed <path> | total | version'
           )
         end