+++ spec/commands/command_factory_spec.rb
@@ -13,7 +13,8 @@ RSpec.describe SimpleCovMcp::Commands::CommandFactory do
         ['summary', SimpleCovMcp::Commands::SummaryCommand],
         ['raw', SimpleCovMcp::Commands::RawCommand],
         ['uncovered', SimpleCovMcp::Commands::UncoveredCommand],
-        ['detailed', SimpleCovMcp::Commands::DetailedCommand]
+        ['detailed', SimpleCovMcp::Commands::DetailedCommand],
+        ['total', SimpleCovMcp::Commands::TotalCommand]
       ].each do |command_name, command_class|
         it "creates a #{command_class.name.split('::').last} for \"#{command_name}\"" do
           command = described_class.create(command_name, cli_context)
@@ -27,7 +28,7 @@ RSpec.describe SimpleCovMcp::Commands::CommandFactory do
         [
           'invalid_cmd',
           'invalid command',
-          /list | summary <path> | raw <path> | uncovered <path> | detailed <path> | version/
+          /list | summary <path> | raw <path> | uncovered <path> | detailed <path> | total | version/
         ],
         [nil,           'nil command',            nil],
         ['',            'empty string command',   nil],
@@ -47,7 +48,7 @@ RSpec.describe SimpleCovMcp::Commands::CommandFactory do
       commands = described_class.available_commands
       expect(commands).to be_an(Array)
       expect(commands).to contain_exactly('list', 'version', 'summary', 'raw', 'uncovered',
-        'detailed')
+        'detailed', 'total')
     end
 
     it 'returns the keys from COMMAND_MAP' do
@@ -67,6 +68,7 @@ RSpec.describe SimpleCovMcp::Commands::CommandFactory do
       expect(described_class::COMMAND_MAP['raw']).to eq(SimpleCovMcp::Commands::RawCommand)
       expect(described_class::COMMAND_MAP['uncovered']).to eq(SimpleCovMcp::Commands::UncoveredCommand)
       expect(described_class::COMMAND_MAP['detailed']).to eq(SimpleCovMcp::Commands::DetailedCommand)
+      expect(described_class::COMMAND_MAP['total']).to eq(SimpleCovMcp::Commands::TotalCommand)
     end
   end
 end